requirements.txt
llama-index
openai
python-dotenv
pypdf # For handling PDF files
python-docx # For handling .docx files
pydantic # For data validation and structured output
# Add any other specific libraries you might need, e.g., pandas for data manipulation
pandas
python-dotenv
